Add ToDoItem class
Added ToDoItem class that has the following attributes:
    unique_id
    title
    created_at
    completed_at
    task_status

Unique_id is a string representing unique hex values.
Title will act as the title of the task.
Create_at will contain a date in which the task was created.
Completed_at will contain the date in which the task has been marked complete.
Task status will represent whether the task is in progress or complete.

There are getters for all attributes which are public except
completed_at as access to that attribute is limited and should be only
shown when the task state is complete.

There is also support for changing the status from in progress to
complete which will update the completed_at attribute.

Added method to return a string representing the status of the
task_status enum.

Support for the << operator so that the class can be used with
std::cout.