Update Date class
Overloaded ==, !=, and = operators so that they can be used in the
==, !=, and = operators of the ToDoItem class.

The overloaded == operator will return true if all attributes are the
same, and the != operator will return true if any of them are different.

The = operator will assign all value of another class to the class
instance calling the operator method.