Update ToDoItem class
Added functionality for the ==, !=, = operators.

The == operator will compare all attributes with another class instance
and if they are the same, the method will return true. Otherwise it will
return false.

The reverse is true for the != operator, if any of the attributes are
not the same, then the operator method call will return true. Otherwise
it will return false.

The = operator will assign all attributes of the input to the class
instance that called the operator method